game/fmv: split into platform-specific files
Extract the Emscripten FMV implementation (HTML5 <video> via EM_JS)
into fmv_emscripten.c, leaving fmv.c as desktop-only (FFmpeg VIDEO
API). Wire them up via meson conditional compilation.

Also remove redundant outer #ifdef EMSCRIPTEN_BUILD guards from
audio_sample_emscripten.c, audio_stream_emscripten.c, and
image_emscripten.c since they are already conditionally compiled.
